lint(gui): move API client out of its lonely dir
